python_web/products/pyproject.toml
[project]
name = "products"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.*"
dependencies = [
    "backend-common",
    "fastapi>=0.129.0",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.13.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "gunicorn>=23.0.0",
    "aiokafka>=0.10.0",
    "elasticsearch[async]>=8.0.0",
    "ruff>=0.15.9",
]

[tool.uv.sources]
backend-common = { path = "../../shared", editable = true }

[tool.ruff.lint]
ignore = ["E712", "E402"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
